Move file finder tests to their own file
Previously, the size of the file finder file would cause code formatting to time out.
