+++ typos.toml
@@ -10,7 +10,7 @@ extend-exclude = [
     # Vim makes heavy use of partial typing tables
     "crates/vim/*",
     # Editor and file finder rely on partial typing and custom in-string syntax
-    "crates/file_finder/src/file_finder.rs",
+    "crates/file_finder/src/file_finder_tests.rs",
     "crates/editor/src/editor_tests.rs",
     # :/
     "crates/collab/migrations/20231009181554_add_release_channel_to_rooms.sql",
